add edit and delete icon to subcategory and fix test
